requirements.txt
python-dotenv>=1.0.0
mcp[cli]>=1.4.0
httpx>=0.27.0
aiofiles>=23.2.1
dulwich>=0.22.0,<1.0.0
